curl: disable C source code generator
The --libcurl command line option causes curl to generate C source
code which, when compiled and linked with libcurl, creates a binary
which behaves in the same way as curl when run with the other options
passed on the curl command line.

  https://curl.haxx.se/docs/manpage.html#--libcurl

It's a development tool and not generally useful on the target,
especially if the target doesn't contain a toolchain etc.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Signed-off-by: Ross Burton <ross.burton@intel.com>